@@ -15,16 +15,18 @@
  */
 package com.github.ziplet.filter.compression;
 
-import com.mockrunner.mock.web.MockHttpServletRequest;
-import com.mockrunner.mock.web.MockServletContext;
-import com.mockrunner.mock.web.WebMockObjectFactory;
-import com.mockrunner.servlet.ServletTestModule;
+
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.mock.web.MockFilterChain;
+import org.springframework.mock.web.MockFilterConfig;
+import org.springframework.mock.web.MockHttpServletRequest;
+import org.springframework.mock.web.MockHttpServletResponse;
+
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.util.Random;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 /**
  * Can be used to generate load on {@link CompressingFilter}.
@@ -37,15 +39,13 @@ public final class LoadRunner {
         // do nothing
     }
 
-    public static void main(String... args) {
+    public static void main(String... args) throws ServletException, IOException {
 
-        WebMockObjectFactory factory = new WebMockObjectFactory();
-        MockServletContext context = factory.getMockServletContext();
-        context.setInitParameter("debug", "true");
-        context.setInitParameter("statsEnabled", "true");
-        ServletTestModule module = new ServletTestModule(factory);
-        module.addFilter(new CompressingFilter(), true);
-        module.setDoChain(true);
+        MockFilterConfig context = new MockFilterConfig();
+        context.addInitParameter("debug", "true");
+        context.addInitParameter("statsEnabled", "true");
+        CompressingFilter filter = new CompressingFilter();
+        filter.init(context);
 
         Random r = new Random(0xDEADBEEFL);
         final String[] data = new String[200];
@@ -55,29 +55,29 @@ public final class LoadRunner {
             data[i] = new String(bytes);
         }
 
-        module.setServlet(new HttpServlet() {
+        HttpTestServlet servlet = new HttpTestServlet() {
             @Override
             public void doGet(HttpServletRequest request,
-                HttpServletResponse response) throws IOException {
+                              HttpServletResponse response) throws IOException {
                 PrintWriter writer = response.getWriter();
                 for (String string : data) {
                     writer.print(string);
                 }
             }
-        });
-        MockHttpServletRequest request = factory.getMockRequest();
+        };
+        MockHttpServletRequest request = new MockHttpServletRequest();
         request.addHeader("Accept-Encoding", "gzip");
 
         long start = System.currentTimeMillis();
         int iterations = 1000;
         for (int i = 0; i < iterations; i++) {
-            module.doGet();
+            new MockFilterChain(servlet, filter).doFilter(request, new MockHttpServletResponse());
         }
         long end = System.currentTimeMillis();
         long time = end - start;
         System.out
-            .println(
-                "Completed in " + time + "ms (" + (double) time / iterations + " per request)");
+                .println(
+                        "Completed in " + time + "ms (" + (double) time / iterations + " per request)");
 
     }
 }
