@@ -17,9 +17,10 @@ package com.github.ziplet.filter.compression;
 
 import com.github.ziplet.filter.compression.statistics.CompressingFilterStats;
 import com.github.ziplet.filter.compression.statistics.CompressingFilterStatsImpl;
+import junit.framework.TestCase;
+
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
-import junit.framework.TestCase;
 
 /**
  * Tests {@link StatsInputStream}.
@@ -37,7 +38,7 @@ public final class StatsInputStreamTest extends TestCase {
         super.setUp();
         bais = new ByteArrayInputStream(new byte[100]);
         statsIn = new MockStatsInputStream(bais, new CompressingFilterStatsImpl(),
-            StatsField.REQUEST_INPUT_BYTES);
+                StatsField.REQUEST_INPUT_BYTES);
     }
 
     public void testStats() throws Exception {
@@ -53,7 +54,7 @@ public final class StatsInputStreamTest extends TestCase {
     private static final class MockStatsInputStream extends StatsInputStream {
 
         public MockStatsInputStream(InputStream inputStream, CompressingFilterStats stats,
-            StatsField field) {
+                                    StatsField field) {
             super(inputStream, stats, field);
         }
 
