@@ -17,9 +17,10 @@ package com.github.ziplet.filter.compression;
 
 import com.github.ziplet.filter.compression.statistics.CompressingFilterStats;
 import com.github.ziplet.filter.compression.statistics.CompressingFilterStatsImpl;
+import junit.framework.TestCase;
+
 import java.io.ByteArrayOutputStream;
 import java.io.OutputStream;
-import junit.framework.TestCase;
 
 /**
  * Tests {@link StatsOutputStream}.
@@ -36,7 +37,7 @@ public final class StatsOutputStreamTest extends TestCase {
         super.setUp();
         baos = new ByteArrayOutputStream();
         statsOut = new MockStatsOutputStream(baos, new CompressingFilterStatsImpl(),
-            StatsField.RESPONSE_INPUT_BYTES);
+                StatsField.RESPONSE_INPUT_BYTES);
     }
 
     public void testStats() throws Exception {
@@ -61,7 +62,7 @@ public final class StatsOutputStreamTest extends TestCase {
     private static final class MockStatsOutputStream extends StatsOutputStream {
 
         public MockStatsOutputStream(OutputStream outputStream, CompressingFilterStats stats,
-            StatsField field) {
+                                     StatsField field) {
             super(outputStream, stats, field);
         }
 
